+++ sqlstore/src/main/resources/db/migration/V0530__auto_main.sql
@@ -1,0 +1,6 @@
+-- increase_groupId_length
+
+ALTER TABLE DetailedSample CHANGE COLUMN groupId groupId VARCHAR(100) DEFAULT NULL;
+ALTER TABLE DetailedLibrary CHANGE COLUMN groupId groupId VARCHAR(100) DEFAULT NULL;
+
+